HUB75_P10_3535_16X32_4S

+++ src/main.cpp
@@ -575,7 +575,7 @@ void LoadUsbPackageSizeMultiplier() {
 #ifdef ZEDMD_HD_HALF
 void SaveYOffset() {
   File f = LittleFS.open("/y_offset.val", "w");
-  f.write(yOffset);
+  f.write(static_cast<uint8_t>(yOffset));
   f.close();
 }
 
@@ -585,7 +585,7 @@ void LoadYOffset() {
     SaveYOffset();
     return;
   }
-  yOffset = f.read();
+  yOffset = static_cast<int8_t>(f.read());
   f.close();
 }
 #endif
